Move AccessorSynthesis out to its own file
